+++ src/App.js
@@ -1,23 +1,12 @@
-import logo from './logo.svg';
-import './App.css';
-
+import React from 'react';
+import './style/App.css';
+import ControlBar from './js/views/components/block/control-bar.js';
 function App() {
-  return (
+  return(
     <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
+      <div className="header"></div>
+      <div className="contents"></div>
+      <ControlBar/>
     </div>
   );
 }
